fix: satisfy pkg-file-count limits after main merge
Consolidate durable session handlers into handlers_factory.go, move durable
service wiring into runtime_sessions.go, and relocate runtime-backed API
loopback tests under pkg/api/servertests.
